Edits to `conftest` and utility functions location
